# Chronos: support numpy in and out onnx/openvino models without converting to tensors