--- a/python/chronos/src/bigdl/chronos/forecaster/base_forecaster.py
+++ b/python/chronos/src/bigdl/chronos/forecaster/base_forecaster.py
@@ -762,7 +762,9 @@ class BasePytorchForecaster(Forecaster):
                         self.accelerated_model.eval()
                         yhat = _pytorch_fashion_inference(model=self.accelerated_model,
                                                           input_data=data,
-                                                          batch_size=batch_size)
+                                                          batch_size=batch_size,
+                                                          output_tensor=\
+                                                            self.optimized_model_output_tensor)
             if not is_local_data:
                 yhat = np_to_xshard(yhat, self.workers_per_node, prefix="prediction")
             return yhat
@@ -834,7 +836,8 @@ class BasePytorchForecaster(Forecaster):
                                       "please call .quantize() method first")
                 return _pytorch_fashion_inference(model=self.accelerated_model,
                                                   input_data=data,
-                                                  batch_size=batch_size)
+                                                  batch_size=batch_size,
+                                                  output_tensor=self.optimized_model_output_tensor)
             else:
                 if self.accelerate_method != "onnxruntime_fp32":
                     self.build_onnx()
@@ -842,7 +845,8 @@ class BasePytorchForecaster(Forecaster):
                                                          self.thread_num)
                 return _pytorch_fashion_inference(model=self.accelerated_model,
                                                   input_data=data,
-                                                  batch_size=batch_size)
+                                                  batch_size=batch_size,
+                                                  output_tensor=self.optimized_model_output_tensor)
 
     def predict_with_openvino(self, data, batch_size=32, quantize=False):
         """
@@ -912,7 +916,8 @@ class BasePytorchForecaster(Forecaster):
                                       "please call .quantize() method first")
                 return _pytorch_fashion_inference(model=self.accelerated_model,
                                                   input_data=data,
-                                                  batch_size=batch_size)
+                                                  batch_size=batch_size,
+                                                  output_tensor=self.optimized_model_output_tensor)
             else:
                 if self.accelerate_method != "openvino_fp32":
                     self.build_openvino()
@@ -920,7 +925,8 @@ class BasePytorchForecaster(Forecaster):
                                                          self.thread_num)
                 return _pytorch_fashion_inference(model=self.accelerated_model,
                                                   input_data=data,
-                                                  batch_size=batch_size)
+                                                  batch_size=batch_size,
+                                                  output_tensor=self.optimized_model_output_tensor)
 
     def predict_with_jit(self, data, batch_size=32, quantize=False):
         """
@@ -1119,7 +1125,8 @@ class BasePytorchForecaster(Forecaster):
                         self.accelerated_model.eval()
                         yhat = _pytorch_fashion_inference(model=self.accelerated_model,
                                                           input_data=input_data,
-                                                          batch_size=batch_size)
+                                                          batch_size=batch_size,
+                                                          output_tensor=self.optimized_model_output_tensor)
 
             aggregate = 'mean' if multioutput == 'uniform_average' else None
             return Evaluator.evaluate(self.metrics, target,
@@ -1211,6 +1218,7 @@ class BasePytorchForecaster(Forecaster):
                 target = np.concatenate(tuple(val[1] for val in data), axis=0)
         else:
             input_data, target = data
+
         if not self.context_enabled:
             self.cxt_manager = ForecasterContextManager(self, self.thread_num, True)
         else:
@@ -1229,7 +1237,8 @@ class BasePytorchForecaster(Forecaster):
                     self.build_onnx()
                 yhat = _pytorch_fashion_inference(model=self.accelerated_model,
                                                   input_data=input_data,
-                                                  batch_size=batch_size)
+                                                  batch_size=batch_size,
+                                                  output_tensor=self.optimized_model_output_tensor)
 
         aggregate = 'mean' if multioutput == 'uniform_average' else None
         return Evaluator.evaluate(self.metrics, target, yhat, aggregate=aggregate)
@@ -1556,6 +1565,7 @@ class BasePytorchForecaster(Forecaster):
                                                           accelerator="onnxruntime",
                                                           onnxruntime_session_options=sess_options,
                                                           output_tensors=False)
+        self.optimized_model_output_tensor = False
         self.accelerate_method = "onnxruntime_fp32"
         self.optimized_model_thread_num = thread_num
 
@@ -1595,6 +1605,7 @@ class BasePytorchForecaster(Forecaster):
                                                           accelerator="openvino",
                                                           thread_num=thread_num,
                                                           output_tensors=False)
+        self.optimized_model_output_tensor = False
         self.accelerate_method = "openvino_fp32"
         self.optimized_model_thread_num = thread_num
 
@@ -1966,10 +1977,12 @@ class BasePytorchForecaster(Forecaster):
         if accelerator == 'onnxruntime':
             q_model.output_tensors = False
             self.accelerated_model = q_model
+            self.optimized_model_output_tensor = False
             self.accelerate_method = "onnxruntime_int8"
         if accelerator == 'openvino':
             q_model.output_tensors = False
             self.accelerated_model = q_model
+            self.optimized_model_output_tensor = False
             self.accelerate_method = "openvino_int8"
         if accelerator is None:
             self.accelerated_model = q_model