--- a/python/chronos/src/bigdl/chronos/forecaster/nbeats_forecaster.py
+++ b/python/chronos/src/bigdl/chronos/forecaster/nbeats_forecaster.py
@@ -168,6 +168,7 @@ class NBeatsForecaster(BasePytorchForecaster):
         self.quantize_available = True
         self.checkpoint_callback = True
         self.use_hpo = True
+        self.optimized_model_output_tensor = True
 
         super().__init__()