--- a/python/chronos/src/bigdl/chronos/pytorch/utils.py
+++ b/python/chronos/src/bigdl/chronos/pytorch/utils.py
@@ -81,7 +81,8 @@ def _pytorch_fashion_inference(model, input_data, batch_size=None, output_tensor
     else:
         # Reduce time of type transforms between tensor and numpy
         if isinstance(input_data, DataLoader):
-            input_sample_list = list(map(lambda x: x[0].numpy(), input_data))
+            input_sample = tuple(batch[0].numpy() for batch in input_data)
+            input_sample_list = [np.concatenate(input_sample, axis=0)]
         elif isinstance(input_data, np.ndarray):
             input_sample_list = [input_data]
         else: